added collectottransaction

+++ src/main/java/com/sample/cash/management/system/model/Request/AddHotelRequest.java
@@ -15,5 +15,5 @@ public class AddHotelRequest {
     private String emailAddress;
     private String password;
     private String collectorEmail;
-    private Long balance;
+    private double balance;
 }
